Separate package for helper methods
