+++ entrypoint/airgap/airgap_suite_test.go
@@ -24,7 +24,7 @@ func TestMain(m *testing.M) {
 	flag.Var(&flags.Destroy, "destroy", "Destroy cluster after test")
 	flag.StringVar(&flags.AirgapFlag.RegistryUsername, "registryUsername", "", "private registry username")
 	flag.StringVar(&flags.AirgapFlag.RegistryPassword, "registryPassword", "", "private registry password")
-	flag.StringVar(&flags.AirgapFlag.TarballType, "tarballType", "", "artifact tarball typr")
+	flag.StringVar(&flags.AirgapFlag.TarballType, "tarballType", "", "artifact tarball type")
 	flag.Parse()
 
 	_, err := config.AddEnv()
